requirements.txt
millify
streamlit_extras
plotly
altair
sqlite3